[#2640] ImageryLayerCatalogItem: Move setClockCurrentTime() out into its own file so that it can be accessed from multiple locations.
